requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn
python-dotenv
django-import-export
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0
